refactor(modding): move systems topology logic out of the aggregate system type
