+++ OpenSolarMax.Game/Modding/SortedSystemTypeCollection.cs
@@ -1,0 +1,9 @@
+namespace OpenSolarMax.Game.Modding;
+
+internal record ImmutableSortedSystemTypeCollection(
+    ImmutableSortedSystemTypes Input,
+    ImmutableSortedSystemTypes Ai,
+    ImmutableSortedSystemTypes Simulate,
+    ImmutableSortedSystemTypes Render,
+    ImmutableSortedSystemTypes Preview
+);
